refactor: make behaviors to policy companion object
